dbeaver/pro#2515 fixes
